Rewrite how config_attribs are handled for pulses
Expose config file attributes in one place and one place only both for
numerical and analytical pulses: the `config_attribs` attribute. This
avoids a lot of confusion with multiple levels of data potentially
conflicting with each other, and genrally streamlines the interface.

Also, distinguish complex and real pulses by the type of the amplitude,
instead of a 'mode' flag.